<feature>[identity]: move share resource logic to ShareResourceHelper
Autowired class ShareResourceHelper used to:
* Share resources to specific accounts
* Share resources to public
* Revoke resources sharing to specific accounts
* Revoke resources sharing to public

Related: ZSV-6755
Related: ZSV-6559
